requirements.txt
python-telegram-bot == 12.8
requests == 2.24.0
